p4i integration Integrate 128313 from 3 to main: Compile under Windows!
[git-p4: depot-paths = "//depot/qt/main/": change = 128314]

+++ qmake/generators/mac/pbuilder_pbx.cpp
@@ -340,7 +340,7 @@ ProjectBuilderMakefileGenerator::writeMakeParts(QTextStream &t)
 	    files = fileFixify(files);
 	    for(QStringList::Iterator file_it = files.begin(); file_it != files.end(); ++file_it) {
 		QString file = (*file_it);
-		if(file.length() >= 2 && (file[0] == '"' || file[0] == '\'') && file[file.length()-1] == file[0])
+		if(file.length() >= 2 && (file[0] == '"' || file[0] == '\'') && file[(int) file.length()-1] == file[0])
 		    file = file.mid(1, file.length()-2);
 		if(file.endsWith(Option::cpp_moc_ext) || file.endsWith(Option::prl_ext)) 
 		    continue;
@@ -550,7 +550,7 @@ ProjectBuilderMakefileGenerator::writeMakeParts(QTextStream &t)
 	    for(QStringList::Iterator it = tmp.begin(); it != tmp.end();) {
 		bool remove = FALSE;
 		QString library, name, opt = (*it).trimmed();
-		if(opt.length() >= 2 && (opt[0] == '"' || opt[0] == '\'') && opt[opt.length()-1] == opt[0])
+		if(opt.length() >= 2 && (opt[0] == '"' || opt[0] == '\'') && opt[(int) opt.length()-1] == opt[0])
 		    opt = opt.mid(1, opt.length()-2);
 		if(opt.startsWith("-L")) {
 		    QString r = opt.right(opt.length() - 2);
